Moved random fns to random.ts
