--- src/utils/get-random-element.ts
@@ -1,3 +1,0 @@
-const getRandomElement = (list: any[]) => list[Math.floor(Math.random() * list.length)]
-
-export default getRandomElement;
